Fixed not using pick_valid to refine some encroachments
In theory, if the point that was encroaching was obtained with Pick valid,
the facet should be refined using pick_valid.
In practice, it doesn't have any noticable impact on the mesh, but it costs
a lot of time to use pick_valid.
Leaving the theory version in code for correctness, for now.