Deformationless aniso geodesic CVT
Relatively minor change: there were 2 functions that I used to check how
good the centroids were from deformation-based AGCVT in the base case of
an isotropic (or aniso uniform) metric field. The first function uses the
(sum of ci ti / sum of ti) formula and the other one the polygon centroid
formula with the Voronoi vertices.

In both cases, we can use an anisotropic area computation instead and it works
surprisingly well...