Fixed sometimes creating orphans while refining a geodesic Voronoi diagram
Very rarely, one could create orphan if we use local spreading from
a new seed. In theory this can't happen but in practice we only have
an approximation of the geodesic distance.

We add 'children' to each canvas vertex, that is the other canvas vertices
that use this given vertex as ancestor. When an ancestor changes values,
all the children are invalidated. There are two different ways to rebuild
the children, not sure which method is best yet...

Only 2D code here.