Fixed a bug in Delaunay local's refinement algorithm
Tedious bug: when inserting a new point p, an older star can be modified
by the insertion of p but this insertion of p can also make older
points become in conflict with the newly modified older star...

In practice this changes... pretty much nothing; it hardly ever occured.
However, the cost of the fix is not null since it results in querying
(possibly multiple times) the kd tree for all the stars modified
by the insertion of a new point.