Revert "Simplified typedef"
This reverts commit fecfaacbaf78e0455f08e1c213e024228846a561.

Although this typedef is still useless, it's useful to check the sanity of
the names of template parameters of the periodic traits classes. If some
template paremeters are named like typedefs of the kernel (which is valid code,
but not ideal) some compilers will bug and fail to compile.

See also commit cbde1087088cbb1e5c52af49a9bbdbdd628d4b0e