Fix crash on escape again:
Something in the dock widgets of camera_positions_plugin and Parameterization_plugin
mess up the order of destruction of the viewers and the scene, so I re-introduced the property "is_destroyed",
only if the viewers are destroyed by the scene's destructor. That should avoid the crash on windows AND fix the crash on closure.