Invalid (X)HTMLsequence
In packages.html we get:
`Element style is not declared in p list of possible children`
as the `style` tag is only allowed in the `head` part of a file.

The specific changes are only needed for this one file, so it has been implemented using the post processing facility of te CGAL documentation generation.