Correction of (X)HTML warnings in case of Advanced topics
Based on #3713 a  number of problems were solved, but also some new issues introduced.
- BaseDoxyfile.h see to it that the CGALAdvancedBegin is in a detailed part, so be sure to terminate the brief part by an extra carriage return
- commands like \pre, \note read till the beginning of the next section but a \htmlonly is not seen as a section separator, so insert an extra carriage return
- some places  missed an "Advanced" indicator, used the cgalAdvancedType here.

##Note
in Periodic_2_triangulation_2.h there appear to be a number of images that are a bit in thin air (Captions: "Insertion of a point on an edge." and "Insertion in a face."). Also just above these images theer are a number of functions that have documentation, but this is non-doxygen documentation and as such not seen by doxygen